Reproduction of bug relating to adding "track by $index" to ng-repeated items.
Results of the sort can be inconsistent and incorrect when the items listed were generated by ng-repeat utilizing "track by"

In the updated demo (demo/groups.html):
When a user sorts position 3 to position 1, the result will be position 2.
Position 4 to position 1 will result in the final position being 4.

This applies to all items root level and nested.

It's hard to nail down an exact case to reproduce, however, repeated and broad sorting eventually ends up more and more inconsistent as the steps increase.